+++ app/src/main/java/eu/kanade/test/DummyTracker.kt
@@ -16,6 +16,7 @@ data class DummyTracker(
     override val name: String,
     override val supportsReadingDates: Boolean = false,
     override val supportsPrivateTracking: Boolean = false,
+    override val supportsNovels: Boolean = false,
     override val isLoggedIn: Boolean = false,
     override val isLoggedInFlow: Flow<Boolean> = flowOf(false),
     val valLogo: Int = R.drawable.brand_anilist,